Rename to vpdq_test_util to avoid pytest action
